Move parsers in `cardano-cli.hs` into `Cardano.CLI.Parsers`
